// packages/web/src/generated/rust-types/DocCommentOutput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Doc comment output
 */
export type DocCommentOutput = { id: string, docId: string, userId: string, userName: string, commentKey: string, content: string, isResolved: boolean, parentId: string | null, replies: Array<DocCommentOutput>, createdAt: string, updatedAt: string, };